+++ src/main/java/com/bewitchment/client/render/tile/TileRenderCauldron.java
@@ -1,7 +1,12 @@
 package com.bewitchment.client.render.tile;
 
+import java.util.Optional;
+
+import org.lwjgl.opengl.GL11;
+
 import com.bewitchment.client.ResourceLocations;
 import com.bewitchment.common.tile.TileCauldron;
+
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.OpenGlHelper;
@@ -16,9 +21,6 @@ import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fluids.Fluid;
 import net.minecraftforge.fluids.FluidRegistry;
 import net.minecraftforge.fluids.FluidStack;
-import org.lwjgl.opengl.GL11;
-
-import java.util.Optional;
 
 /**
  * This class was created by Arekkuusu on 09/03/2017.
@@ -35,7 +37,7 @@ public class TileRenderCauldron extends TileEntitySpecialRenderer<TileCauldron> 
 			FluidStack fluidStack = optional.get();
 			Fluid fluid = fluidStack.getFluid();
 			ResourceLocation location = fluid.getStill();
-			double level = (double) fluidStack.amount / (Fluid.BUCKET_VOLUME * 2D);
+			double level = fluidStack.amount / (Fluid.BUCKET_VOLUME * 2D);
 
 			GlStateManager.pushMatrix();
 			GlStateManager.disableLighting();
